requirements.txt
django>=3.0, <3.1
djangorestframework>=3.11, <3.12
psycopg2>=2.7.5, <2.8.0
Pillow>=6.2.0, <6.3.0

flake8>=3.6.0, <3.7.0
